+++ src/test/java/com/iexec/worker/feign/CustomCoreFeignClientTests.java
@@ -1,0 +1,153 @@
+/*
+ * Copyright 2023 IEXEC BLOCKCHAIN TECH
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.iexec.worker.feign;
+
+import com.iexec.common.notification.TaskNotificationType;
+import com.iexec.common.replicate.ReplicateStatus;
+import com.iexec.common.replicate.ReplicateStatusUpdate;
+import com.iexec.worker.feign.client.CoreClient;
+import feign.FeignException;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Test;
+import org.junit.jupiter.params.ParameterizedTest;
+import org.junit.jupiter.params.provider.EnumSource;
+import org.mockito.InjectMocks;
+import org.mockito.Mock;
+import org.mockito.MockitoAnnotations;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+
+import java.util.List;
+
+import static org.assertj.core.api.Assertions.assertThat;
+import static org.mockito.Mockito.*;
+
+class CustomCoreFeignClientTests {
+    private static final String AUTHORIZATION = "authorization";
+    private static final String CHAIN_TASK_ID = "0x123";
+
+    @Mock
+    private LoginService loginService;
+    @Mock
+    private CoreClient coreClient;
+    @InjectMocks
+    private CustomCoreFeignClient customCoreFeignClient;
+
+    @BeforeEach
+    void init() {
+        MockitoAnnotations.openMocks(this);
+    }
+
+    //region getCoreVersion
+    @Test
+    void shouldGetCoreVersion() {
+        when(coreClient.getCoreVersion()).thenReturn(ResponseEntity.ok("X.Y.Z"));
+        assertThat(customCoreFeignClient.getCoreVersion()).isEqualTo("X.Y.Z");
+        verifyNoInteractions(loginService);
+    }
+
+    @Test
+    void shouldNotGetCoreVersionWhenError() {
+        when(coreClient.getCoreVersion()).thenThrow(FeignException.class);
+        assertThat(customCoreFeignClient.getCoreVersion()).isNull();
+        verifyNoInteractions(loginService);
+    }
+    //endregion
+
+    //region getComputingTasks
+    @Test
+    void shouldGetComputingTasks() {
+        when(loginService.getToken()).thenReturn(AUTHORIZATION);
+        when(coreClient.getComputingTasks(AUTHORIZATION)).thenReturn(ResponseEntity.ok(List.of(CHAIN_TASK_ID)));
+        final List<String> tasks = customCoreFeignClient.getComputingTasks();
+        assertThat(tasks).containsExactly(CHAIN_TASK_ID);
+        verify(loginService, never()).login();
+    }
+
+    @Test
+    void shouldNotGetComputingTasksWhenBadLogin() {
+        when(loginService.getToken()).thenReturn(AUTHORIZATION);
+        when(loginService.login()).thenReturn(AUTHORIZATION);
+        when(coreClient.getComputingTasks(AUTHORIZATION)).thenThrow(FeignException.Unauthorized.class);
+        final List<String> tasks = customCoreFeignClient.getComputingTasks();
+        assertThat(tasks).isEmpty();
+        verify(loginService, times(3)).login();
+    }
+
+    @Test
+    void shouldNotGetComputingTasksWhenError() {
+        when(loginService.getToken()).thenReturn(AUTHORIZATION);
+        when(coreClient.getComputingTasks(AUTHORIZATION)).thenThrow(FeignException.class);
+        final List<String> tasks = customCoreFeignClient.getComputingTasks();
+        assertThat(tasks).isEmpty();
+        verify(loginService, never()).login();
+    }
+    //endregion
+
+    // region updateReplicateStatus
+    @ParameterizedTest
+    @EnumSource(value = HttpStatus.class, names = {"OK", "ALREADY_REPORTED"})
+    void shouldUpdateReplicatesStatus(HttpStatus status) {
+        final ReplicateStatusUpdate statusUpdate = ReplicateStatusUpdate
+                .builder()
+                .status(ReplicateStatus.COMPLETING)
+                .build();
+
+        when(loginService.getToken()).thenReturn(AUTHORIZATION);
+        when(coreClient.updateReplicateStatus(AUTHORIZATION, CHAIN_TASK_ID, statusUpdate))
+                .thenReturn(ResponseEntity.status(status).body(TaskNotificationType.PLEASE_CONTINUE));
+
+        final TaskNotificationType nextAction = customCoreFeignClient.updateReplicateStatus(CHAIN_TASK_ID, statusUpdate);
+
+        assertThat(nextAction).isEqualTo(TaskNotificationType.PLEASE_CONTINUE);
+    }
+
+    @Test
+    void shouldNotUpdateReplicatesStatusWhenBadLogin() {
+        final ReplicateStatusUpdate statusUpdate = ReplicateStatusUpdate
+                .builder()
+                .status(ReplicateStatus.COMPLETING)
+                .build();
+
+        when(loginService.getToken()).thenReturn(AUTHORIZATION);
+        when(coreClient.updateReplicateStatus(AUTHORIZATION, CHAIN_TASK_ID, statusUpdate))
+                .thenThrow(FeignException.Unauthorized.class);
+
+        final TaskNotificationType nextAction = customCoreFeignClient.updateReplicateStatus(CHAIN_TASK_ID, statusUpdate);
+
+        assertThat(nextAction).isNull();
+        verify(loginService).login();
+    }
+
+    @Test
+    void shouldNotUpdateReplicatesStatusWhenError() {
+        final ReplicateStatusUpdate statusUpdate = ReplicateStatusUpdate
+                .builder()
+                .status(ReplicateStatus.COMPLETING)
+                .build();
+
+        when(loginService.getToken()).thenReturn(AUTHORIZATION);
+        when(coreClient.updateReplicateStatus(AUTHORIZATION, CHAIN_TASK_ID, statusUpdate))
+                .thenThrow(FeignException.Forbidden.class);
+
+        final TaskNotificationType nextAction = customCoreFeignClient.updateReplicateStatus(CHAIN_TASK_ID, statusUpdate);
+
+        assertThat(nextAction).isNull();
+        verify(loginService, never()).login();
+    }
+    // endregion
+}